165: Temporary Messaging Removed
